Fixed merge conflicts, additional cleanup
