+++ src/client/setup/setup-base.tsx
@@ -3,7 +3,7 @@ import { Outlet } from "react-router-dom";
 import { ChessboardWrapper } from "../chessboard/chessboard-wrapper";
 import { PropsWithChildren, ReactNode } from "react";
 import { ChessEngine } from "../../common/chess-engine";
-import { Side } from "../../common/types";
+import { Side } from "../../common/game-types";
 
 interface SetupBaseProps extends PropsWithChildren {
     actions?: ReactNode;
