+++ src/common/client-types.ts
@@ -1,0 +1,21 @@
+export enum GameType {
+    COMPUTER = "computer",
+    HUMAN = "human",
+} /**
+ * Defines a type of session.
+ */
+
+export enum ClientType {
+    HOST = "host",
+    CLIENT = "client",
+    SPECTATOR = "spectate",
+} /*
+ * Difficulties available for AI.
+ */
+
+export enum Difficulty {
+    BABY = 0,
+    BEGINNER = 1,
+    INTERMEDIATE = 2,
+    ADVANCED = 3,
+}
